# Conflicts:
#	Elegant.Web/Views/Product/GetProductById.cshtml
#	Elegant.Web/Views/Shared/Header.cshtml

+++ Elegant.DAL/Storages/DbProductsStorage.cs
@@ -1,8 +1,9 @@
-﻿using Elegant.Business.Services;
-using Elegant.Core.Models;
+﻿using Elegant.Core.Models;
+using Elegant.DAL.Interfaces;
 using Microsoft.EntityFrameworkCore;
 
 namespace Elegant.DAL.Storages;
+
 public class DbProductsStorage : IProductsStorage
 {
     private readonly EfCoreDbContext _dbContext;
@@ -12,39 +13,59 @@ public class DbProductsStorage : IProductsStorage
         _dbContext = efCoreDbContext;
     }
 
-    public List<Product> GetAll() => _dbContext
+    public async Task<List<Product>> GetAllAsync(CancellationToken cancellationToken) =>await _dbContext
         .Products
         .Include(product => product.CartItems)
         .Include(product => product.ImageItems)
-        .ToList();
-    public void Add(Product product)
+        .AsSplitQuery()
+        .ToListAsync(cancellationToken: cancellationToken);
+
+    public async Task AddAsync(Product product, CancellationToken cancellationToken)
     {
         _dbContext.Products.Add(product);
-        _dbContext.SaveChanges();
+        await _dbContext.SaveChangesAsync(cancellationToken);
     }
-    public void Update(Product product)
+
+    public async Task UpdateAsync(Product product, CancellationToken cancellationToken)
     {
-        var productInStorage = GetById(product.Id);
-        productInStorage.Name = product.Name;
-        productInStorage.Cost = product.Cost;
-        productInStorage.Description = product.Description;
-        productInStorage.ImagePath = product.ImagePath;
-        _dbContext.SaveChanges();
+        var productInStorage = await GetByIdAsync(product.Id, cancellationToken);
+        if (productInStorage != null)
+        {
+            productInStorage.Name = product.Name;
+            productInStorage.Cost = product.Cost;
+            productInStorage.Description = product.Description;
+            productInStorage.ImageItems = product.ImageItems;
+        }
+
+        await _dbContext.SaveChangesAsync(cancellationToken);
     }
-    public List<Product> Search(string name)
+
+    public async Task<List<Product>> GetByNameAsync(string name, CancellationToken cancellationToken)
     {
-        if (string.IsNullOrWhiteSpace(name))
-            return new List<Product>();
-        return _dbContext.Products.Where(product => product.Name.ToLower().StartsWith(name.ToLower())).ToList();
+        return string.IsNullOrWhiteSpace(name)
+            ? new List<Product>()
+            : await _dbContext.Products.Where(product => product.Name.ToLower().StartsWith(name.ToLower()))
+                .ToListAsync(cancellationToken);
     }
-    public void Remove(Guid productId)
+
+    public async Task RemoveAsync(Guid productId, CancellationToken cancellationToken)
     {
-        var product = GetById(productId);
+        var product = await GetByIdAsync(productId, cancellationToken);
+        if (product == null)
+        {
+            return;
+        }
+
         _dbContext.Products.Remove(product);
-        _dbContext.SaveChanges();
+        await _dbContext.SaveChangesAsync(cancellationToken);
     }
-    public Product GetById(Guid id)
+
+    public async Task<Product?> GetByIdAsync(Guid id, CancellationToken cancellationToken)
     {
-        return _dbContext.Products.Include(product => product.CartItems).Include(product => product.ImageItems).FirstOrDefault(pr => pr.Id == id);
+        return await _dbContext.Products
+            .Include(product => product.CartItems)
+            .Include(product => product.ImageItems)
+            .AsSplitQuery()
+            .FirstOrDefaultAsync(pr => pr.Id == id, cancellationToken);
     }
 }
